move CreateProfileView into Binders package
